+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Introduced Rector to modernize code !3084 @\_fridtjof_
 - Update backend dependencies !3094 @\_fridtjof_
 - Modernize code to use PHP 8.1 features !3095 @\_fridtjof_
+- Modernized BlogControl to a Symfony controller !3096 @\_fridtjof_
 
 ## Development documentation